+++ Examples/Examples-tvOS/Pods/Target Support Files/IGListKit/IGListKit-umbrella.h
@@ -1,0 +1,37 @@
+#ifdef __OBJC__
+#import <UIKit/UIKit.h>
+#endif
+
+#import "IGListAdapter.h"
+#import "IGListAdapterDataSource.h"
+#import "IGListAdapterDelegate.h"
+#import "IGListAdapterUpdater.h"
+#import "IGListAdapterUpdaterDelegate.h"
+#import "IGListAssert.h"
+#import "IGListBatchUpdateData.h"
+#import "IGListCollectionContext.h"
+#import "IGListCollectionView.h"
+#import "IGListDiff.h"
+#import "IGListDiffable.h"
+#import "IGListDisplayDelegate.h"
+#import "IGListExperiments.h"
+#import "IGListIndexPathResult.h"
+#import "IGListIndexSetResult.h"
+#import "IGListKit.h"
+#import "IGListMacros.h"
+#import "IGListMoveIndex.h"
+#import "IGListMoveIndexPath.h"
+#import "IGListReloadDataUpdater.h"
+#import "IGListScrollDelegate.h"
+#import "IGListSectionController.h"
+#import "IGListSectionType.h"
+#import "IGListSingleSectionController.h"
+#import "IGListStackedSectionController.h"
+#import "IGListSupplementaryViewSource.h"
+#import "IGListUpdatingDelegate.h"
+#import "IGListWorkingRangeDelegate.h"
+#import "NSObject+IGListDiffable.h"
+
+FOUNDATION_EXPORT double IGListKitVersionNumber;
+FOUNDATION_EXPORT const unsigned char IGListKitVersionString[];
+