+++ Source/IGListCollectionContext.h
@@ -15,40 +15,42 @@ NS_ASSUME_NONNULL_BEGIN
 @protocol IGListSectionType;
 
 /**
- The collection context provides limited access to the collection related information that
- section controllers need for things like sizing, dequeing cells, insterting/deleting/reloading, etc.
+ The collection context provides limited access to the collection-related information that
+ section controllers need for operations like sizing, dequeing cells, insterting, deleting, reloading, etc.
  */
 @protocol IGListCollectionContext <NSObject>
 
 /**
- Size of the collection view. Provided primarily for sizing cells.
+ The size of the collection view. You may use this for sizing cells.
  */
 @property (nonatomic, readonly) CGSize containerSize;
 
 /**
- Query for the index a cell in the collection relative to the section controller.
+ Returns the index of the specified cell in the collection relative to the section controller.
 
  @param cell              An existing cell in the collection.
  @param sectionController The section controller requesting this information.
 
- @return The index of the cell or NSNotFound if it does not exist in the collection.
+ @return The index of the cell or `NSNotFound` if it does not exist in the collection.
  */
-- (NSUInteger)indexForCell:(UICollectionViewCell *)cell
+- (NSInteger)indexForCell:(UICollectionViewCell *)cell
          sectionController:(IGListSectionController<IGListSectionType> *)sectionController;
 
 /**
- Query for a cell in the collection. May return nil if the cell is offscreen.
+ Returns the cell in the collection at the specified index for the section controller.
 
  @param index             The index of the desired cell.
  @param sectionController The section controller requesting this information.
 
  @return The collection view cell, or `nil` if not found.
+ 
+ @warning This method may return `nil` if the cell is offscreen.
  */
 - (nullable __kindof UICollectionViewCell *)cellForItemAtIndex:(NSInteger)index
                                              sectionController:(IGListSectionController<IGListSectionType> *)sectionController;
 
 /**
- Query for the visible cells for the given section controller.
+ Returns the visible cells for the given section controller.
 
  @param sectionController The section controller requesting this information.
 
@@ -68,22 +70,22 @@ NS_ASSUME_NONNULL_BEGIN
                    animated:(BOOL)animated;
 
 /**
- Query the section index of an section controller.
+ Returns the section index of an section controller.
 
  @param sectionController An section controller object.
 
- @return The section index of the list if found, otherwise `NSNotFound`.
+ @return The section index of the controller if found, otherwise `NSNotFound`.
  */
-- (NSUInteger)sectionForSectionController:(IGListSectionController<IGListSectionType> *)sectionController;
+- (NSInteger)sectionForSectionController:(IGListSectionController<IGListSectionType> *)sectionController;
 
 /**
- Dequeues a cell from the UICollectionView reuse pool.
+ Dequeues a cell from the collection view reuse pool.
 
  @param cellClass         The class of the cell you want to dequeue.
  @param sectionController The section controller requesting this information.
  @param index             The index of the cell.
 
- @return A cell dequeued from the reuse pool or newly created.
+ @return A cell dequeued from the reuse pool or a newly created one.
 
  @note This method uses a string representation of the cell class as the identifier.
  */
@@ -92,14 +94,14 @@ NS_ASSUME_NONNULL_BEGIN
                                                       atIndex:(NSInteger)index;
 
 /**
- Dequeues a cell from the UICollectionView reuse pool.
+ Dequeues a cell from the collection view reuse pool.
  
  @param nibName           The name of the nib file.
- @param bundle            The bundle in which to search for the nib file. If nil, this method looks for the nib file in the main bundle.
+ @param bundle            The bundle in which to search for the nib file. If `nil`, this method searches the main bundle.
  @param sectionController The section controller requesting this information.
  @param index             The index of the cell.
  
- @return A cell dequeued from the reuse pool or newly created.
+ @return A cell dequeued from the reuse pool or a newly created one.
  
  @note This method uses a string representation of the cell class as the identifier.
  */
@@ -109,27 +111,27 @@ NS_ASSUME_NONNULL_BEGIN
                                                           atIndex:(NSInteger)index;
 
 /**
- Dequeues a storyboard prototype cell from the UICollectionView reuse pool.
+ Dequeues a storyboard prototype cell from the collection view reuse pool.
  
  @param identifier        The identifier of the cell prototype in storyboard.
  @param sectionController The section controller requesting this information.
  @param index             The index of the cell.
  
- @return A cell dequeued from the reuse pool or newly created.
+ @return A cell dequeued from the reuse pool or a newly created one.
  */
 - (__kindof UICollectionViewCell *)dequeueReusableCellFromStoryboardWithIdentifier:(NSString *)identifier
                                                               forSectionController:(IGListSectionController <IGListSectionType> *)sectionController
                                                                            atIndex:(NSInteger)index;
 
 /**
- Dequeues a supplementary view from the UICollectionView reuse pool.
+ Dequeues a supplementary view from the collection view reuse pool.
 
  @param elementKind       The kind of supplementary veiw.
  @param sectionController The section controller requesting this information.
  @param viewClass         The class of the supplementary view.
  @param index             The index of the supplementary vew.
 
- @return A supplementary view dequeued from the reuse pool or newly created.
+ @return A supplementary view dequeued from the reuse pool or a newly created one.
 
  @note This method uses a string representation of the view class as the identifier.
  */
@@ -139,29 +141,29 @@ NS_ASSUME_NONNULL_BEGIN
                                                                       atIndex:(NSInteger)index;
 
 /**
- Dequeues a supplementary view from the UICollectionView reuse pool.
+ Dequeues a supplementary view from the collection view reuse pool.
  
  @param elementKind       The kind of supplementary veiw.
  @param identifier        The identifier of the supplementary view in storyboard.
  @param sectionController The section controller requesting this information.
  @param index             The index of the supplementary vew.
  
- @return A supplementary view dequeued from the reuse pool or newly created.
+ @return A supplementary view dequeued from the reuse pool or a newly created one.
  */
 - (__kindof UICollectionReusableView *)dequeueReusableSupplementaryViewFromStoryboardOfKind:(NSString *)elementKind
                                                                              withIdentifier:(NSString *)identifier
                                                                        forSectionController:(IGListSectionController<IGListSectionType> *)sectionController
                                                                                     atIndex:(NSInteger)index;
 /**
- Dequeues a supplementary view from the UICollectionView reuse pool.
+ Dequeues a supplementary view from the collection view reuse pool.
  
  @param elementKind       The kind of supplementary veiw.
  @param sectionController The section controller requesting this information.
  @param nibName           The name of the nib file.
- @param bundle            The bundle in which to search for the nib file. If nil, this method looks for the nib file in the main bundle.
+ @param bundle            The bundle in which to search for the nib file. If `nil`, this method searches the main bundle.
  @param index             The index of the supplementary vew.
  
- @return A supplementary view dequeued from the reuse pool or newly created.
+ @return A supplementary view dequeued from the reuse pool or a newly created one.
  
  @note This method uses a string representation of the view class as the identifier.
  */
@@ -181,7 +183,7 @@ NS_ASSUME_NONNULL_BEGIN
                         atIndexes:(NSIndexSet *)indexes;
 
 /**
- Inserts cells in the feed.
+ Inserts cells in the list.
 
  @param sectionController The section controller who's cells need inserting.
  @param indexes           The indexes of items that need inserting.
@@ -190,7 +192,7 @@ NS_ASSUME_NONNULL_BEGIN
                         atIndexes:(NSIndexSet *)indexes;
 
 /**
- Deletes cells in the feed.
+ Deletes cells in the list.
 
  @param sectionController The section controller who's cells need deleted.
  @param indexes           The indexes of items that need deleting.
@@ -199,14 +201,14 @@ NS_ASSUME_NONNULL_BEGIN
                         atIndexes:(NSIndexSet *)indexes;
 
 /**
- Reload the entire section controller.
+ Reloads the entire section controller.
 
- @param sectionController The list object who's cells need reloading.
+ @param sectionController The section controller who's cells need reloading.
  */
 - (void)reloadSectionController:(IGListSectionController<IGListSectionType> *)sectionController;
 
 /**
- Batch many cell-level updates in a single transaction.
+ Batches and performs many cell-level updates in a single transaction.
 
  @param animated   A flag indicating if the transition should be animated.
  @param updates    A block containing all of the cell updates.
@@ -215,9 +217,10 @@ NS_ASSUME_NONNULL_BEGIN
  @note Use this method to batch cell updates (inserts, deletes, reloads) into a single transaction. This lets you
  make many changes to your data store and perform all the transitions at once.
 
- For example, inside your section controllers, you may want to delete /and/ insert into the data source that backs your
- section controller:
+ For example, inside your section controllers, you may want to delete *and* insert into the data source that backs your
+ section controller. For example:
 
+ ```
  [self.collectionContext performBatchItemUpdates:^{
  // perform data source changes inside the update block
  [self.items addObject:newItem];
@@ -229,8 +232,9 @@ NS_ASSUME_NONNULL_BEGIN
  NSIndexSet *deletes = [NSIndexSet indexSetWithIndex:0];
  [self.collectionContext deleteInSectionController:self deletes];
  } completion:nil];
+ ```
 
- Note that you **must** perform data modifications **inside** the update block. Updates will not be performed
+ @warning You **must** perform data modifications **inside** the update block. Updates will not be performed
  synchronously, so you should make sure that your data source changes only when necessary.
  */
 - (void)performBatchAnimated:(BOOL)animated updates:(void (^)())updates completion:(nullable void (^)(BOOL finished))completion;